fastapi_app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
cryptography==41.0.5
Pillow==10.1.0
numpy==1.26.2
pycryptodome==3.19.0
python-jose==3.3.0
python-dotenv==1.0.0 